fpaths, handling

+++ queries/coordination/file_handling.py
@@ -2,28 +2,31 @@ from pathlib import Path
 
 import duckdb
 from huggingface_hub import HfFileSystem
+from tqdm import tqdm
 
 
-def initialize_crawler(pattern: str):
+def initialize_crawler(pattern: str, skip_existing=True):
     fs = HfFileSystem()
     files = fs.glob(pattern)
     sub_name = "_".join(pattern.split("datasets/")[1].split("/")[0:-1])
 
     data_path = Path(f"data/{sub_name}")
-    data_path.mkdir(parents=True, exist_ok=True)
-    to_crawl_path = data_path / "to_crawl.txt"
-    seen_path = data_path / "seen.txt"
-    error_path = data_path / "error.txt"
 
-    with open(to_crawl_path, "w") as f:
-        for file in files:
-            f.write(f"{file}\n")
+    if skip_existing or not data_path.exists():
+        data_path.mkdir(parents=True, exist_ok=True)
+        to_crawl_path = data_path / "to_crawl.txt"
+        seen_path = data_path / "seen.txt"
+        error_path = data_path / "error.txt"
 
-    with open(seen_path, "w") as f:
-        pass
+        with open(to_crawl_path, "w") as f:
+            for file in files:
+                f.write(f"{file}\n")
 
-    with open(error_path, "w") as f:
-        pass
+        with open(seen_path, "w") as f:
+            pass
+
+        with open(error_path, "w") as f:
+            pass
 
     return data_path
 
@@ -48,7 +51,7 @@ def crawl(data_path, query_path, crawl_errors=False):
     else:
         to_crawl = list(set(error) - set(seen))
 
-    for file in to_crawl:
+    for file in tqdm(to_crawl):
         file = file.strip()
         fname = file.split("/")[-1]
         try:
